+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.0.1 under development
 
+- New #63: Allow pass one listener to muiltiple event class names in `ListenerCollection::add()` (vjik)
 - Bug #62: Add a support listeners with union type of event (vjik)
 
 ## 1.0.0 February 16, 2021