+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## 1.0.1 under development
 
-- New #63: Allow pass one listener to muiltiple event class names in `ListenerCollection::add()` (vjik)
-- Bug #62: Add a support listeners with union type of event (vjik)
+- New #63: Allow adding a single listener to multiple event class names at once in `ListenerCollection::add()` (vjik)
+- Bug #62: Support listeners for events defined with union types (vjik)
 
 ## 1.0.0 February 16, 2021