+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # Yii Event Dispatcher Change Log
 
-## 1.0.1 under development
+## 1.0.1 July 28, 2021
 
-- New #63: Allow adding a single listener to multiple event class names at once in `ListenerCollection::add()` (vjik)
 - Bug #62: Support listeners for events defined with union types (vjik)
+- New #63: Allow adding a single listener to multiple event class names at once in `ListenerCollection::add()` (vjik)
 
 ## 1.0.0 February 16, 2021